[C-1664] Improve lineup, app-tab, rewards perf
