+++ packages/mobile/src/screens/notifications-screen/NotificationsScreen.tsx
@@ -6,7 +6,7 @@ import { useDispatch } from 'react-redux'
 
 import IconNotification from 'app/assets/images/iconNotification.svg'
 import { Screen, ScreenContent, ScreenHeader } from 'app/components/core'
-import { usePopToTopOnDrawerOpen } from 'app/hooks/usePopToTopOnDrawerOpen'
+import { useAppTabScreen } from 'app/hooks/useAppTabScreen'
 
 import { NotificationList } from './NotificationList'
 const { markAllAsViewed } = notificationsActions
@@ -16,7 +16,7 @@ const messages = {
 }
 
 export const NotificationsScreen = () => {
-  usePopToTopOnDrawerOpen()
+  useAppTabScreen()
   const dispatch = useDispatch()
 
   useFocusEffect(
